Added --repair and --last options to CLI
Two new options added to the CLI:

The --repair option calls repair_lastrun() and attempts to fix any
issues with the lastRotation variable in lastrun.py by extrapolating
the rotation data of prior sales.

The --last (-l) option is a simple way to pass the most recent sale
link (which is stored in lastrun.py anyways) to get_content as the
testLink argument.